# openapi.yaml
openapi: 3.0.3
info:
  title: AfroGoBackend API
  version: "1.0.0"
  description: >
    AfroGoBackend v1 - Headless API for logistics, routing, payouts, billing,
    and ops variance review. Use this OpenAPI skeleton as the contract for v1.
servers:
  - url: https://api-dev.afrogo.co.za/v1
    description: Dev
paths:
...
  /billing/invoices:
    post:
      tags: [Billing]
      summary: Generate invoice
      operationId: generateInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceSummary'
  /billing/invoices/{invoiceId}:
    get:
      tags: [Billing]
      summary: Get invoice metadata
      operationId: getInvoice
      parameters:
        - in: path
          name: invoiceId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          description: Not found
  /billing/payment-webhook:
    post:
      tags: [Billing]
      summary: Payment provider webhook
      operationId: paymentWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWebhookPayload'
      responses:
        '200':
          description: Processed
        '401':
          description: Signature invalid
...
components:
  schemas:
...
    CreateInvoiceRequest:
      type: object
      required: [merchantId, items]
      properties:
        merchantId:
          type: string
        dueDate:
          type: string
          format: date
        currency:
          type: string
        customerEmail:
          type: string
          format: email
        customerName:
          type: string
        items:
          type: array
          items:
            type: object
            required: [description, qty, unitPrice]
            properties:
              description:
                type: string
              qty:
                type: number
              unitPrice:
                type: number
        metadata:
          type: object
          additionalProperties: true
    InvoiceSummary:
      type: object
      properties:
        invoiceId:
          type: string
        totalAmount:
          type: number
        currency:
          type: string
    Invoice:
      allOf:
        - $ref: '#/components/schemas/InvoiceSummary'
        - type: object
          properties:
            merchantId:
              type: string
            status:
              type: string
            s3Key:
              type: string
    PaymentWebhookPayload:
      type: object
      required: [invoiceId, status, amount]
      properties:
        invoiceId:
          type: string
        status:
          type: string
        amount:
          type: number
        providerReference:
          type: string
        signature:
          type: string
        data:
          type: object
          additionalProperties: true
...